+++ src/components/PermissionsManagement/RolePermissions.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { permissions } from '../../utils/permissions';
-import { permissionLabel } from './PermissionsManagement';
+import { permissionLabel } from './UserRoleTab';
+import { Button } from 'reactstrap';
 
 function getKeyByValue(object, value) {
     return Object.keys(object).find(key => object[key] === value);
@@ -24,16 +25,18 @@ const mapPermissionToLabel = (role) => {
 function RolePermissions(props) {
     const permissions = mapPermissionToLabel(props.role);
 
+    const permissionsCopy = permissions;
+
+    console.log('user role permissions: ', permissionsCopy);
+
     return (
         <>
             <h2>{props.header}</h2>
             {props.permissionsList.map(permission => (
                 <p style={{color: permissions.includes(permission) ? 'green' : 'red'}}>
-                    {permission} {permissions.includes(permission) ? '-' : '+'}
+                    {permission} {permissions.includes(permission) ? <Button color='danger'>-</Button> : <Button color='success'>+</Button>}
                 </p>
-            ))}
-                
-            
+            ))}        
         </>
     );
 }
